% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arred.R
\name{arred}
\alias{arred}
\title{Formata números para impressão}
\usage{
arred(x, digitos = 3, dm = ",", bm = "")
}
\arguments{
\item{x}{Número a ser formatado}

\item{digitos}{Quantidade de dígitos}

\item{dm}{Separador de decimal}

\item{bm}{Separador de milhar}
}
\value{
Número formatado
}
\description{
Formata números para impressão
}
\examples{
arred(25.8549, digitos = 2)

}
