% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_cochranq.R
\name{eta_cochranq}
\alias{eta_cochranq}
\title{Maximum-corrected measure of effect size eta - Para o teste Q de Cochran}
\usage{
eta_cochranq(teste)
}
\arguments{
\item{teste}{O resultado do teste Q de Cochran feito pela função rstatix::cochran_qtest}
}
\description{
Maximum-corrected measure of effect size eta - Para o teste Q de Cochran
}
