% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_linear.R
\name{forest_linear}
\alias{forest_linear}
\title{Cria um gráfico de floresta para a regressão linear}
\usage{
forest_linear(mod)
}
\arguments{
\item{mod}{Modelo de regressão linear (criado pela função lm)}
}
\description{
Cria um gráfico de floresta para a regressão linear
}
