% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_obs_esp_ader.R
\name{graf_obs_esp_ader}
\alias{graf_obs_esp_ader}
\title{Gráfico dos valores observados e esperados para o qui-quadrado de aderência}
\usage{
graf_obs_esp_ader(
  dados,
  v1,
  props = "",
  xlab = "",
  cor = c("grey75", "grey45")
)
}
\arguments{
\item{dados}{Banco de dados}

\item{v1}{Variável categórica}

\item{props}{Proporções esperadas das categorias}

\item{xlab}{Título eixo x}

\item{cor}{Cor das barras}
}
\value{

}
\description{
Gráfico dos valores observados e esperados para o qui-quadrado de aderência
}
