% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_obs_esp_ind.R
\name{graf_obs_esp_ind}
\alias{graf_obs_esp_ind}
\title{Gráfico dos valores observados e esperados para o qui-quadrado de independência}
\usage{
graf_obs_esp_ind(
  dados,
  v1,
  v2,
  cat_graf_v2,
  xlab = "",
  cor = c("grey75", "grey45"),
  MacDonald = F
)
}
\arguments{
\item{dados}{Banco de dados}

\item{v1}{Variável categórica 1}

\item{v2}{Variável categórica 2}

\item{cat_graf_v2}{Categoria da variável 2 que será representada no gráfico}

\item{xlab}{Título do eixo x (referente à variável 1)}

\item{cor}{Cor das barras}

\item{MacDonald}{Usar correção de MacDonald para os resíduos?}
}
\value{

}
\description{
Gráfico dos valores observados e esperados para o qui-quadrado de independência
}
