% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_obs_esp_ind_2.R
\name{graf_obs_esp_ind_2}
\alias{graf_obs_esp_ind_2}
\title{Gráfico dos valores observados e esperados para o qui-quadrado de independência, mostrando todas as categorias}
\usage{
graf_obs_esp_ind_2(
  dados,
  v1,
  v2,
  xlab = "",
  cor = c("grey75", "grey45"),
  asteriscos = T,
  MacDonald = F
)
}
\arguments{
\item{dados}{Banco de dados}

\item{v1}{Variável categórica 1}

\item{v2}{Variável categórica 2}

\item{xlab}{Título do eixo x}

\item{cor}{Cor das barras}

\item{asteriscos}{Incluir asteriscos no gráfico?}

\item{MacDonald}{Usar correção de MacDonald para os resíduos?}
}
\value{

}
\description{
Gráfico dos valores observados e esperados para o qui-quadrado de independência, mostrando todas as categorias
}
