% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graf_obs_esp_ind_3.R
\name{graf_obs_esp_ind_3}
\alias{graf_obs_esp_ind_3}
\title{Gráfico dos valores observados e esperados para o qui-quadrado de independência, com barras vazias}
\usage{
graf_obs_esp_ind_3(
  dados,
  v1,
  v2,
  xlab = "",
  filllab = "",
  cor = c("grey75", "grey45"),
  MacDonald = F
)
}
\arguments{
\item{dados}{Banco de dados}

\item{v1}{Variável categórica 1}

\item{v2}{Variável categórica 2}

\item{xlab}{Título do eixo x (v1)}

\item{filllab}{Título do preechimento (v2)}

\item{cor}{Cores das barras}

\item{MacDonald}{Usar correção de MacDonald para os resíduos?}
}
\value{

}
\description{
Gráfico dos valores observados e esperados para o qui-quadrado de independência, com barras vazias
}
