% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_val.R
\name{p_val}
\alias{p_val}
\title{Formata o valor de p}
\usage{
p_val(x)
}
\arguments{
\item{x}{Valor de p a ser formatado}

\item{idioma}{Caso "PT", a vírgula passa a ser o separador de decimal}
}
\value{
Valor de p formatado
}
\description{
Formata o valor de p
}
\examples{
p_val(0.00876)
p_val(0.0004)

}
