% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_cochranq.R
\name{r_cochranq}
\alias{r_cochranq}
\title{Chance corrected measure of effect size r - Para o teste Q de Cochran}
\usage{
r_cochranq(df)
}
\arguments{
\item{df}{Banco de dados no formato wide e excluindo o ID}
}
\description{
Chance corrected measure of effect size r - Para o teste Q de Cochran
}
