% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_linear.R
\name{reg_linear}
\alias{reg_linear}
\title{Tabela e gráfico para regressão linear}
\usage{
reg_linear(mod, add_global_p = F)
}
\arguments{
\item{add_global_p}{}
}
\description{
Tabela e gráfico para regressão linear
}
